Create version-specific buildRequest hooks
