+++ src/execute/oas3/style-serializer.js
