+++ test/execute/main.js
@@ -1,6 +1,6 @@
 import expect, {createSpy, spyOn} from 'expect'
 import xmock from 'xmock'
-import {execute, buildRequest, baseUrl, applySecurities, self as stubs} from '../../src/execute'
+import {execute, buildRequest, baseUrl, self as stubs} from '../../src/execute'
 
 // Supported shape...  { spec, operationId, parameters, securities, fetch }
 // One can use operationId or pathItem + method
@@ -924,172 +924,6 @@ describe('execute', () => {
     const req = fetchSpy.calls[0].arguments[0]
     expect(fetchSpy.calls.length).toEqual(1)
     expect(fetchSpy.calls[0].arguments[0].body).toBeA(FormData)
-  })
-
-  describe('applySecurities', function () {
-    it('should NOT add any securities, if the operation does not require it', function () {
-      const spec = {
-        host: 'swagger.io',
-        basePath: '/v1',
-        security: [{apiKey: []}],
-        paths: {
-          '/one': {
-            get: {
-              operationId: 'getMe'
-            }
-          }
-        },
-        securityDefinitions: {
-          apiKey: {
-            in: 'header',
-            name: 'api_key',
-            type: 'apiKey'
-          }
-        }
-      }
-
-      const securities = {}
-      const request = {
-        url: 'http://swagger.io/v1/one',
-        method: 'GET'
-      }
-
-      const applySecurity = applySecurities({request, securities, operation: spec.paths['/one'].get, spec})
-
-      expect(applySecurity).toEqual({
-        url: 'http://swagger.io/v1/one',
-        method: 'GET'
-      })
-    })
-
-    it('should add a basic auth if operation requires it and has header passed', function () {
-      const spec = {
-        host: 'swagger.io',
-        basePath: '/v1',
-        security: [{authMe: []}],
-        paths: {
-          '/one': {
-            get: {
-              operationId: 'getMe',
-              security: [{authMe: []}]
-            }
-          }
-        },
-        securityDefinitions: {
-          authMe: {
-            type: 'basic'
-          }
-        }
-      }
-
-      const request = {
-        url: 'http://swagger.io/v1/one',
-        method: 'GET',
-        query: {}
-      }
-      const securities = {
-        authorized: {
-          authMe: {
-            header: 'Basic Zm9vOmJhcg=='
-          }
-        }
-      }
-
-      const applySecurity = applySecurities({request, securities, operation: spec.paths['/one'].get, spec})
-
-      expect(applySecurity.headers).toEqual({
-        authorization: 'Basic Zm9vOmJhcg=='
-      })
-    })
-
-    it('should add a basic auth if operation requires it', function () {
-      const spec = {
-        host: 'swagger.io',
-        basePath: '/v1',
-        security: [{authMe: []}],
-        paths: {
-          '/one': {
-            get: {
-              operationId: 'getMe',
-              security: [{authMe: []}]
-            }
-          }
-        },
-        securityDefinitions: {
-          authMe: {
-            type: 'basic'
-          }
-        }
-      }
-
-      const request = {
-        url: 'http://swagger.io/v1/one',
-        method: 'GET',
-        query: {}
-      }
-      const securities = {
-        authorized: {
-          authMe: {
-            username: 'foo',
-            password: 'bar'
-          }
-        }
-      }
-
-      const applySecurity = applySecurities({request, securities, operation: spec.paths['/one'].get, spec})
-
-      expect(applySecurity.headers).toEqual({
-        authorization: 'Basic Zm9vOmJhcg=='
-      })
-    })
-
-    it('should be able to apply multiple auths', function () {
-      const spec = {
-        host: 'swagger.io',
-        basePath: '/v1',
-        security: [{authMe: []}, {apiKey: []}],
-        paths: {
-          '/one': {
-            get: {
-              operationId: 'getMe',
-              security: [{authMe: []}, {apiKey: []}]
-            }
-          }
-        },
-        securityDefinitions: {
-          authMe: {
-            type: 'basic'
-          },
-          apiKey: {
-            in: 'header',
-            name: 'api_key',
-            type: 'apiKey'
-          }
-        }
-      }
-
-      const request = {
-        url: 'http://swagger.io/v1/one',
-        method: 'GET',
-        query: {}
-      }
-      const securities = {
-        authorized: {
-          authMe: {
-            username: 'foo',
-            password: 'bar'
-          },
-          apiKey: 'hello'
-        }
-      }
-
-      const applySecurity = applySecurities({request, securities, operation: spec.paths['/one'].get, spec})
-
-      expect(applySecurity.headers).toEqual({
-        authorization: 'Basic Zm9vOmJhcg==',
-        api_key: 'hello'
-      })
-    })
   })
 
   describe('parameterBuilders', function () {
